requirements.txt
dotenv
livekit-agents[openai,silero,deepgram,cartesia,turn-detector]